fastapi_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
SQLAlchemy
